+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- Minor `--help` output changes
+
 ## [0.2.2] - 2025.01.31
 
 ### Fixed